pyproject.toml
[project]
name = "due-diligence-reporter-mcp"
version = "0.1.0"
description = "MCP server for Alpha School due diligence report generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.188.0",
    "google-auth>=2.47.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.14.1",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.scripts]
due-diligence-reporter-mcp = "due_diligence_reporter.server:main"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.13.1",
    "types-requests>=2.32.4.20260107",
]

[tool.uv]
package = true
